+++ src/Tools/sunem_init.m
@@ -51,21 +51,7 @@ function Const = sunem_init(Const, yVectors)
            Const.runMoMsolver = false;
         end
     end
-
-    % ================================    
-    try
-        set = false;
-        if (Const.runSUNEMMoMsolver)
-            % It is set in the driver
-            set = true;
-        end
-    catch
-        if (~set)
-            % Solver not activated
-           Const.runSUNEMMoMsolver = false;
-        end
-    end
-
+    
    % ================================    
     try
         set = false;
@@ -273,6 +259,20 @@ function Const = sunem_init(Const, yVectors)
            Const.SUNEMifbmomstrfilename = '';
         end
     end
+    
+    % ================================    
+    try
+        set = false;
+        if (Const.SUNEMmomstrfilename)
+            % It is set in the driver
+            set = true;
+        end
+    catch
+        if (~set)
+            % Solver not activated
+           Const.SUNEMmomstrfilename = '';
+        end
+    end
 
     % =================================
     % Other initialisations:
@@ -298,7 +298,7 @@ function Const = sunem_init(Const, yVectors)
 
     % For numerical integration over a triangular domain (based on DBD2011)
     Const.QUAD_PTS = 6;     % Quadrature rule (6 point is a good default)
-    Const.SING = false;     % If set to true, then singularities are treated in the DBD2011 routine for
+    Const.SING = false;      % If set to true, then singularities are treated in the DBD2011 routine for
                             % filling the Z matrix
 
     % Note:other parameters like Omega, K, Lambda, is dependent on the wavelenght and